pyproject.toml
[project]
name = "shuttle-python"
version = "0.1.1"
description = "Shuttle for Python"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aws-cdk-lib>=2.132.0",
    "boto3>=1.39.3",
    "click>=8.2.1",
    "docker>=7.1.0",
    "ruff>=0.12.2",
    "pydantic>=2.11.7",
    "pathspec>=0.12.1",
    "psycopg2-binary>=2.9.10",
]
authors = [
    { name = "Shuttle Python contributors" }
]
maintainers = [
    { name = "Shuttle" }
]
keywords = ["cloud", "ifc", "deployment", "infrastructure", "data", "devops", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.12",
    "Environment :: Console",
    "Natural Language :: English",
    "Topic :: System :: Software Distribution"
]

[project.urls]
Homepage = "https://shuttle.dev/"
Documentation = "https://docs.shuttle.dev/python/welcome/introduction"
Repository = "https://github.com/shuttle-hq/shuttle-python"

[project.scripts]
shuttle = "shuttle.__main__:app"

[tool.setuptools.packages.find]
include = ["shuttle*"]

[tool.setuptools.package-data]
shuttle_infra = ["Dockerfile"]

[dependency-groups]
dev = [
    "jaraco-test>=5.5.1",
    "pudb>=2025.1",
    "pytest>=8.4.1",
    "testcontainers[localstack]>=4.10.0",
    "zipp>=3.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
exclude = [
    "*/test_*.py",
    "/.github",
    "uv.lock",
    "/user-project",
    ".dockerignore"
]

[tool.hatch.build.targets.wheel]
packages = [
    "shuttle",
    "shuttle_aws",
    "shuttle_common",
    "shuttle_infra",
    "shuttle_runtime",
    "shuttle_task"
]
